fix: add explicit byte cast in ProcessBatchAsync test lambda
Error: Cannot implicitly convert type 'int' to 'byte'
Fix: Cast the result of (input.Span[0] * 2) to byte explicitly

In C#, byte * int = int, so an explicit cast is required:
(byte)(input.Span[0] * 2)

+++ tests/HeroCrypt.Tests/PerformanceTests.cs
@@ -374,7 +374,7 @@ public class PerformanceTests
         // Act
         var results = await ParallelCryptoOperations.ProcessBatchAsync<ReadOnlyMemory<byte>, byte>(
             inputs,
-            async input => await Task.FromResult(input.Span[0] * 2));
+            async input => await Task.FromResult((byte)(input.Span[0] * 2)));
 
         // Assert
         Assert.Equal(100, results.Length);
